Improvement and fix to bin/gen-status.sh
The tool now has an -l ls option. This is useful until it can be ported
to GNU ls(1) which does not support the options we need.

Check that the IOCCC_STATUS value is valid. Previously one could have
changed the status to 'foo bar baz' even though it was not a valid
value.