Clarify FAQ 5.5 wrt anonymous authors
The FAQ asks to fix or fill in missing author locations but I have
changed it to say only for non-anonymous authors as only the judges know
exactly what each anonymous author wants. This is added because against
better judgement and making an assumption (and we all know what that
does) I had thought that since location is identifying it must have been
a mistake that location codes were in some of the anonymous author
details (in one case it appeared to not be an error so that one was not
touched).

But the only mistake is that anonymous authors were touched in the first
place so to try and prevent that I have updated the FAQ about not
updating anonymous author details. There might be other places that have
to be updated with this thought and it might be a good idea to have an
FAQ about both past and future handling of anonymous entries but that is
something I will leave to the judges.